+++ LMeter/Helpers/Utils.cs
@@ -96,17 +96,4 @@ namespace LMeter.Helpers
                     "[name_last.1]     =>    L";
         }
     }
-        
-    public enum DrawAnchor
-    {
-        Center = 0,
-        Left = 1,
-        Right = 2,
-        Top = 3,
-        TopLeft = 4,
-        TopRight = 5,
-        Bottom = 6,
-        BottomLeft = 7,
-        BottomRight = 8
-    }
 }